Roll back to H2O's current jsr166y
jsr166 jdk7 - clearly leaves tasks idle on the queues; under a heavy
load (parsing large file on 5 jvms on a 4-cpu machine), some TaskPutKey
will get left hanging - even after all things are idle.  The Task is on
a F/J queue, and the queue threads are idle, parked scanning for work.
jsr166y jdk8 needs a bunch of work to run on with jdk7 libs