Weaver: load Icer classes through currentClassLoader explicitly
javassist loads classes through thread context classloader. But
scalatest sets a child URLClassLoader as the thread context
classloader. So the Class.forName() which immediately follows
the CtClass.toClass() is unable to find the Icer class, because
the current classloader is the parent and threfore unable to find
the Icer class.

While it does not matter for standalone H2O, it is probably a good
practice to explicitly set javassist's classloader to be the
current classloader. This fixes the Mahout scalatest failures too.

Signed-off-by: Anand Avati <avati@redhat.com>