H2O: joinOthers() - method to detect and join existing cloud
This is a helper function in leiu of a real client mode. The idea
is that a server cloud (H2O nodes) is already up and formed, and a
client (e.g Mahout) joins this cloud. The client does not know (and
should not care) how many nodes are in the cloud.

This function is a replacement for a real "connect to cloud" API.

Signed-off-by: Anand Avati <avati@redhat.com>