Remembering potential clients and including them in mutlicast (flatfile mode)
If flatfile is defined, the existing cluster nodes have to inform
client about the cloud shape.

Explanation: In  proper multicast mode, this is ensured
by flooding network with multicast packets. However, in flatfile
mode cluster nodes' heartbeats were not propagated to clients.

Fix: if node receives a heartbeat from client, it extends its internal
list of nodes to multicast (the list is defined by flatfile on command
line) and starts to multicast its heartbeats to the client.

Drawbacks: currently the list of nodes informed by heartbeat is not
modified when a client disappears. Ie., a node which contains a client
in multicast list will continue to send heartbeats to the potentionally
dead client.
This is ok when the client re-appears (even it is a new client)
but it can cause network flooding if there is a high number of dead
clients.