this works for supporting multi-node like h2o1
cd testdir_single_jvm
python test_simple.py -cj ./pytest_config-kevin.json

(the config has my local ip, but the idea is the same as h2o1, the config describes the cloud to be built)

seems to work like h2o1 for copying jars, but it doesn't pass the test if more than 1 jvm is specified.
I think it's because the "cloud_size" in the Cloud.json from h2o-dev is staying = 0. I didn't check whether the h2o stdout/browser says it's clouded up, but the Cloud.json needs to report the right cloud_size (my prior hack was to increment it by 1 for cloud of 1, but that doesn't seem to work here)