Handle OOM during GetKey during NBHM put
NBHM putIfMatch resizing triggered an OOM.
This failed the NBHM call, then the TaskGetKey onAck in the RPC call.
This marked the TGK as completed, so no retries every happened - but the
fetched Key was not installed, the task was not marked "done", the
waiting thread never was woken up (because the Key was not installed),
but no retries happened because the task was completed.
Also, make NBHM use MemoryManager.malloc for large arrays, to handle
OOMs better.