Fix of a race in MRTask. Tasks with modifiable inputs may be resending left and/or right task while local computation is already changing shared data.
Added _modifiesInputs flag to DTask which signals that the task modifies its inputs. If the flag is set and the task has been sent by UDP, RPC will keep the original raw bits and will use those if the task is re-send over the wire instead of re-serializing live task object.