Remove special setChunkSize code in HDFS and NFS file vec. Remove key in FileVec when setting chunk size if not UploadFileVec.
For the UploadFileVec case, this would remove the data, so don't do it
in that case.

This addresses the problem exposed when running in EC2 and reading data
from S3.  The pre-parse step of clearing the caches was actually reading
in all the data instead.  (Whoops.)

(From Brandon with a tweak for the UploadFileVec case.)