PUBDEV-1534: Fix the seed such that GBM weights test passes.
The reason that sometimes it doesn't pass:
Column type affects the histogramming, and column type depends on whether rows are excluded vs just marked with weight 0.

Example:

val weight
1      1
0.5    0
5      1
3.5    0

This Vec will have real-valued type if passed as a whole, but integer weight if
rows with weight 0 are sliced away first.  The resulting histogram will either
be kept at full nbins resolution or potentially shrunk to the discrete integer
range, which will of course affect the split points.