Accidentally mispelled a field while resolving merge conflicts which caused master to fail all of the tests...whoops.
HEXDEV-272 All of the fields of a schema are now stored in the leaf child of the class hierarchy. Changed the implementation of fields() to simply return the fields variable of a schema. The function calls H2O.fail() if it attempts to access a field from a non-leaf child. "response_column" is now moved above "ignored_columns" for every applicable schema. 'own_fields' is also now renamed to 'fields'.