Update TestNG Report feature After running testNG, the report is not as expected, update it with the following things: + Testcase name as input from the spreadsheet file + Mark test with invalid input as INVALID status
This was done by adding a task to gradle and a python script for doing the convert job.

NOTE: Python script was written with XML manipulation. However, the solution was not working due to python xml library adding too much format to the final XML file. So the old-school RE find/replace rocks!