[PUBDEV-1874] Change grid creation behavior.
Now, every call of grid API creates a new grid object
with unique key even it is call of training on the same frame.

TO append models of new grid run to an existing grid, the user has to specify
id of the existing grid. In this case, existing grid object will
be fetched and all generated models will be appended to it.