PUBDEV-2338: Add max_runtime_secs argument for all models. Only in Java/Flow/Scala, not yet in R/Py.
By default, the runtime is unlimited (same as setting to 0), for
max_runtime_secs > 0, EACH (sub)model is stopped when ITS time is up.

Again:

For CV, the timeout applies for each CV-model and also for the main
model. If holdout models are built in parallel (async=T), then that will affect
the amount of work done, of course.

For grid-search, the timeout applies for each model (and it's submodels, if CV)